ディレクトリ整理&型宣言箇所の変更
