+++ client/src/components/Stock/index.tsx
@@ -4,6 +4,17 @@ import StockList from './StockList'
 import StockDetail from './StockDetail'
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom'
 
+export type TStock = {
+  id: number
+  name: string
+  colorNumber: string
+  manufacturingDate: string
+  quantity: number
+  used: boolean
+  storehouse: string
+  file: any
+}
+
 const Stock = () => {
   return (
     <Router>
