+++ client/src/components/Stock/shared/FileUploader/style.tsx
